Add *Actual speed* feature request
When playing videos in really high speed (eg. 100x, yes, with some #solareclipse videos I've done that), it's actually played slower. Maybe my HW is not capable to render video stream that fast or what... I wish to know what was actual playback speed, because it looks it was only ~10x faster.