Update our speed setting when the speed is changed externally.
e.g. Coursera videos have their own speed controls. If the speed
is changed from A to B via their controls, then the video is
paused & resumed, our play callback would reset the speed to A.

We now ensure our speed setting is accurate by updating it in the
ratechange callback, which happens no matter how the speed is
updated.