special-case inject prototype for FB and amazon
Some sites inject overlays over their videos to intercept clicks and
provide own on-screen controls, etc. Unfortunately this makes the VC
controls inaccessible and without any generic workaround... well, short
of relying on keyboard shortcuts.

This is an experiment for special-casing FB and Amazon
- FB we modify stacking context of the parent
- Amazon we inject the controller one level higher, above the overlay

The gotcha here is that this type of behavior is not uncommon, and
special-casing each and every origin like this is not long-term
sustainable. If this sticks.. we'll have to be very selective
about which sites we enable this for.

Closes #210
Closes #134