Affects #626
What has been done?
- Added an event listener for seeked video events and mirrors the
actions of a play event
- Overlay will now update when keypress speed changes are detected
- When the VideoSpeed controller isn't found, the extension will now
look for it in the video element as well as fallback

Why has this been done?
- Kaltura videos act differently in a few cases.  On seek events, a play
event is not emiited. On speed change events, a rateChange event is not
emitted.  Because of this, it causes the overlay not to update on
keypress or seek events.