adjusting default position for x.com mobile
the video speed controller on x.com would appear underneath the back button in fullscreen when using a mobile browser that allows extensions (e.g. kiwi browser). 

This patch uses a domain-specific override to shift the default position down so that it is usable in this specific case. It is no longer stuck underneath the back button.